wallet/wallets/serializers.py
from rest_framework import serializers

from wallets.models import Wallet


class WalletSerializer(serializers.ModelSerializer):
    class Meta:
        model = Wallet
        fields = ("ulid", "balance")
        read_only_fields = ("ulid", "balance")
